Refactor SIDH sources to avoid multiple definitions
When building a shared lib which includes all SIDH types the original structure of
SIDH sources lead to a number of multiple definitions. Refactoring the sources avoids
this issue:

- move some functions to a common/common_funcs.c file. The common functions do not depend
  on fields or parameter which are specific to a SIDH type

- define several fields/arrays/functions as 'static' inside their sources files. Due to
  the sources structure this was possible without affecting functionality. Thus also reduces
  the number of global definitions

- add some #defines to map generic names to SIDH type specific names. This follows the
  technique already in use inside the Pnnn.c files. These #defines were put into the
  Pnnn_internal.h files and thus are accessible to the arithmetic tests.

No changes to the actual implementation of the math-functions.

These changes are currently in effect only when using cmake to build the sources, libs,
and tests